refactor: 공지사항 CQRS 분리 및 도메인 구조 정리
