+++ src/main/java/com/example/Tokkit_server/dto/notices/NoticeResponseDto.java
@@ -1,6 +1,6 @@
 package com.example.Tokkit_server.dto.notices;
 
-import com.example.Tokkit_server.domain.common.Notice;
+import com.example.Tokkit_server.domain.Notice;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
