+++ src/main/java/com/example/Tokkit_server/repository/NoticeRepository.java
@@ -1,6 +1,6 @@
 package com.example.Tokkit_server.repository;
 
-import com.example.Tokkit_server.domain.common.Notice;
+import com.example.Tokkit_server.domain.Notice;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
